refactor: Splits views.py in multiple files
