+++ learning_environment/urls.py
@@ -2,26 +2,29 @@ from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
 
-from . import views
-from .views import SignUpView
-urlpatterns = [
-    # Buildin
-    path("accounts/", include("django.contrib.auth.urls")),
+from learning_environment.views.dashboards import global_dashboard, learner_dashboard
+from learning_environment.views.lessons import TaskListView, LessonDetailView, LessonCreateView, LessonDeleteView
+from learning_environment.views import practice, home, myhome
+from learning_environment.views.users import SignUpView, learner_reset
 
-    # Selfmade
-    path('', views.home, name='home'),
-    path('myhome/', views.myhome, name='myhome'),
-    path('practice/', views.practice, name='practice'),
-    path("signup/", SignUpView.as_view(), name="signup"),
+urlpatterns = [
+    # Build in
+    path('accounts/', include('django.contrib.auth.urls')),
+
+    # Self made
+    path('', home, name='home'),
+    path('myhome/', myhome, name='myhome'),
+    path('practice/', practice, name='practice'),
+    path('signup/', SignUpView.as_view(), name='signup'),
 
     # Backstage
-    path('backstage/tasklist', views.TaskListView.as_view(), name='tasklist'),
-    path('backstage/lesson/<int:pk>', views.LessonDetailView.as_view(), name='lessondetail'),
-    path('backstage/lesson/create', views.LessonCreateView.as_view(), name='lessoncreate'),
-    path('backstage/lesson/delete/<int:pk>', views.LessonDeleteView.as_view(), name='lessondelete'),
-    path('backstage/dashboard/learner', views.learner_dashboard, name='learner_dashboard'),
-    path('backstage/dashboard/global', views.global_dashboard, name='global_dashboard'),
-    path('backstage/reset', views.learner_reset, name='learner_reset'),
+    path('backstage/tasklist', TaskListView.as_view(), name='tasklist'),
+    path('backstage/lesson/<int:pk>', LessonDetailView.as_view(), name='lessondetail'),
+    path('backstage/lesson/create', LessonCreateView.as_view(), name='lessoncreate'),
+    path('backstage/lesson/delete/<int:pk>', LessonDeleteView.as_view(), name='lessondelete'),
+    path('backstage/dashboard/learner', learner_dashboard, name='learner_dashboard'),
+    path('backstage/dashboard/global', global_dashboard, name='global_dashboard'),
+    path('backstage/reset', learner_reset, name='learner_reset'),
 
 ] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
