+++ learning_environment/views/__init__.py
@@ -1,0 +1,2 @@
+from .home import home, myhome
+from .practice import practice
