+++ learning_environment/views/dashboards.py
@@ -1,0 +1,29 @@
+"""
+View elements for different dashboards.
+"""
+from django.shortcuts import render
+
+from learning_environment.models import Solution
+
+
+def learner_dashboard(request):
+    """Prepare data for learner's own dashboard and show it."""
+    solutions = Solution.objects.filter(user=request.user).order_by('timestamp')  # all solutions from current user
+    num_tasks = solutions.count()  # how tasks did this user try
+    correct_solutions = Solution.objects.filter(user=request.user, solved=True).count()  # how many of them were correct?
+    if num_tasks > 0:  # calculate percentage of correct tasks (or 0 if no tasks)
+        tasks_correctness = correct_solutions / num_tasks * 100.0
+    else:
+        tasks_correctness = 0.0
+    return render(request, 'learning_environment/learner_dashboard.html', locals())  # pass all local variable to template
+
+
+def global_dashboard(request):
+    solutions = Solution.objects.all().order_by('timestamp')  # all solutions
+    num_solutions = solutions.count()  # how tasks did this user try
+    correct_solutions = Solution.objects.filter(solved=True).count()  # how many of them were correct?
+    if num_solutions > 0:  # calculate percentage of correct tasks (or 0 if no tasks)
+        tasks_correctness = correct_solutions / num_solutions * 100.0
+    else:
+        tasks_correctness = 0.0
+    return render(request, 'learning_environment/global_dashboard.html', locals())  # pass all local variable to template
